layout-service/requirements.txt
# 1) Install PyTorch for your platform first, e.g. CUDA 12.4:
#    pip install torch --index-url https://download.pytorch.org/whl/cu124
# 2) Then:
#    pip install -r requirements.txt

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pymupdf>=1.24.3
pillow>=10.2.0
numpy>=1.26.0
doclayout-yolo
